refactor(editor): extract arrow, crop, and text state modules (PR 10.1–10.3)
Move existing-arrow editing, crop geometry/drag/apply, and remaining text
size/font/re-edit methods into state child modules without behavior changes.
