+++ docs/CODEBASE_REFACTOR_STATUS.md
@@ -37,7 +37,7 @@ See earlier sections in git history / audit. Short recap:
 
 ### PR 10+: editor state and GTK coordinators — in progress
 
-#### Done (2026-08-05)
+#### Done (2026-08-05 / 2026-08-06)
 
 ##### 1. `EditorState` module tree
 
@@ -45,13 +45,21 @@ See earlier sections in git history / audit. Short recap:
 
 | File | Role | Approx. lines |
 |------|------|--------------:|
-| `mod.rs` | `EditorState` / `TextInputState` structs, shared free helpers, remaining impls | ~2,280 |
-| `text_input.rs` | Text input / fit / commit path | ~580 |
+| `mod.rs` | `EditorState` / `TextInputState` structs, shared free helpers, remaining impls | ~1,330 |
+| `text_input.rs` | Text input/fit/commit + size/font/re-edit/cancel (**PR 10.3**) | ~850 |
+| `crop.rs` | Crop aspect/geometry/drag/apply/fill helpers + `draft_crop_rect` (**PR 10.2**) | ~485 |
+| `arrow.rs` | Existing-arrow style/reverse/control-handle edit (**PR 10.1**) | ~190 |
 | `history.rs` | Undo/redo/push/history availability | ~130 |
-| `drag_draw.rs` | Begin/update/clear/draft/finalize drag + canvas expand | ~480 |
-| `export.rs` | `to_rendered_image` / `to_final_image` / background compose | ~250 |
+| `drag_draw.rs` | Begin/update/clear/draft/finalize drag + canvas expand (new-arrow construction stays here) | ~440 |
+| `export.rs` | `to_rendered_image` / `to_final_image` / background compose; uses `crop::crop_image` | ~250 |
 
-Struct stays in `state/mod.rs` so private field access remains inside the `state` module tree. No new public crate API.
+Struct stays in `state/mod.rs` so private field access remains inside the `state` module tree. No new public crate API. `TextInputState` remains in `mod.rs` (sibling event code consumes it).
+
+**PR 10.1 (2026-08-06):** moved arrow style getters/setters, selected-arrow style mutation, reversal, control-handle hit testing/movement, and finalize/cleanup into `state/arrow.rs` with the two arrow unit tests. Method signatures unchanged.
+
+**PR 10.2 (2026-08-06):** moved crop aspect/ratio, selection init/drag/resize/reset/apply, fill/image helpers, and `draft_crop_rect` into `state/crop.rs`. `crop_image` stays module-private to the state tree via the private `crop` module for `export.rs`.
+
+**PR 10.3 (2026-08-06):** consolidated remaining text methods into `state/text_input.rs` (size/font selection, selected-text lookup, commit/re-edit, test-only update, `cancel_text_edit`). Kept `active_text_edit`/`active_text_entry`/Escape cancel path.
 
 ##### 2. Editor events — first EventContext-owned families
 
@@ -113,7 +121,7 @@ These are the remaining high-coupling coordinators and oversized files. Prefer *
 | `capture/editor/window/mod.rs` | ~4,050 | Dominated by `setup_editor_window_full` + tests |
 | `capture/editor/window/events/mod.rs` | ~3,490 | Dominated by `wire_editor_events` |
 | `capture_overlay.rs` | ~2,430 | **Deferred** process-boundary split (audit: after daemon/capture) |
-| `capture/editor/state/mod.rs` | ~2,280 | More `EditorState` groups still inline (selection, crop, arrow, tool style) |
+| `capture/editor/state/mod.rs` | ~1,330 | More `EditorState` groups still inline (selection, tool style); arrow/crop/text done |
 | `overlay/window/mod.rs` | ~2,060 | Dominated by `setup_window` |
 | `capture/editor/tests.rs` | ~2,050 | Test payload only |
 | `recording/backend.rs` | ~1,930 | Already split out of recording facade; further internal split optional |
@@ -132,8 +140,11 @@ These are the remaining high-coupling coordinators and oversized files. Prefer *
    - Async effects pipeline only if closures already have a clear owner
    - **Defer** `set_draw_func` body and large layout/chrome until Parts ownership is clear
 
-3. **More `EditorState` child modules**
-   - Selection / crop / arrow / tool-style groups still in `state/mod.rs`
+3. **More `EditorState` child modules** (Track A in validation plan)
+   - [x] PR 10.1 arrow editing (`state/arrow.rs`)
+   - [x] PR 10.2 crop (`state/crop.rs`)
+   - [x] PR 10.3 text consolidation (`state/text_input.rs`)
+   - [ ] PR 10.4 selection · PR 10.5 tool-style
    - Keep struct in `state/mod.rs`; child `impl EditorState` only inside the `state` tree
 
 4. **Overlay `setup_window`**
